requirements.txt
Flask==2.1.2
python-dotenv==1.0.0
requests==2.28.2
PyPDF2==3.0.0
pymongo
Werkzeug==2.0.3